requirements.txt
praw
streamlit
openai>=1.0.0
pandas
python-dotenv
requests
